Add library.uefi.sleep_scripts file
Move S3 functionality into sleep_state file
