+++ chipsec/library/uefi/uefi_common.py
@@ -32,7 +32,7 @@ from typing import Dict, List, Tuple, Optional, Any
 from uuid import UUID
 
 from chipsec.library.file import read_file, write_file
-from chipsec.library.logger import logger, dump_buffer, dump_buffer_bytes
+from chipsec.library.logger import logger
 from chipsec.library.defines import bytestostring
 
 # from chipsec.helper.oshelper import helper
@@ -498,350 +498,6 @@ def parse_efivar_file(fname: str, var: Optional[bytes] = None, var_type: int = S
         parse_esal_var(var, var_path)
     else:
         logger().log_warning(f'Unsupported variable type requested: {var_type}')
-
-
-########################################################################################################
-#
-# S3 Resume Boot-Script Parsing Functionality
-#
-########################################################################################################
-
-BOOTSCRIPT_TABLE_OFFSET = 24
-RUNTIME_SCRIPT_TABLE_BASE_OFFSET = 32
-ACPI_VARIABLE_SET_STRUCT_SIZE = 0x48
-S3_BOOTSCRIPT_VARIABLES = ['AcpiGlobalVariable']
-
-MAX_S3_BOOTSCRIPT_ENTRY_LENGTH = 0x200
-
-
-#
-# MdePkg\Include\Pi\PiS3BootScript.h
-#
-# //*******************************************
-# // EFI Boot Script Opcode definitions
-# //*******************************************
-
-class S3BootScriptOpcode:
-    EFI_BOOT_SCRIPT_IO_WRITE_OPCODE = 0x00
-    EFI_BOOT_SCRIPT_IO_READ_WRITE_OPCODE = 0x01
-    EFI_BOOT_SCRIPT_MEM_WRITE_OPCODE = 0x02
-    EFI_BOOT_SCRIPT_MEM_READ_WRITE_OPCODE = 0x03
-    EFI_BOOT_SCRIPT_PCI_CONFIG_WRITE_OPCODE = 0x04
-    EFI_BOOT_SCRIPT_PCI_CONFIG_READ_WRITE_OPCODE = 0x05
-    EFI_BOOT_SCRIPT_SMBUS_EXECUTE_OPCODE = 0x06
-    EFI_BOOT_SCRIPT_STALL_OPCODE = 0x07
-    EFI_BOOT_SCRIPT_DISPATCH_OPCODE = 0x08
-    EFI_BOOT_SCRIPT_TERMINATE_OPCODE = 0xFF
-
-
-class S3BootScriptOpcode_MDE (S3BootScriptOpcode):
-    EFI_BOOT_SCRIPT_DISPATCH_2_OPCODE = 0x09
-    EFI_BOOT_SCRIPT_INFORMATION_OPCODE = 0x0A
-    EFI_BOOT_SCRIPT_PCI_CONFIG2_WRITE_OPCODE = 0x0B
-    EFI_BOOT_SCRIPT_PCI_CONFIG2_READ_WRITE_OPCODE = 0x0C
-    EFI_BOOT_SCRIPT_IO_POLL_OPCODE = 0x0D
-    EFI_BOOT_SCRIPT_MEM_POLL_OPCODE = 0x0E
-    EFI_BOOT_SCRIPT_PCI_CONFIG_POLL_OPCODE = 0x0F
-    EFI_BOOT_SCRIPT_PCI_CONFIG2_POLL_OPCODE = 0x10
-
-#
-# EdkCompatibilityPkg\Foundation\Framework\Include\EfiBootScript.h
-#
-
-
-class S3BootScriptOpcode_EdkCompat (S3BootScriptOpcode):
-    EFI_BOOT_SCRIPT_MEM_POLL_OPCODE = 0x09
-    EFI_BOOT_SCRIPT_INFORMATION_OPCODE = 0x0A
-    EFI_BOOT_SCRIPT_PCI_CONFIG2_WRITE_OPCODE = 0x0B
-    EFI_BOOT_SCRIPT_PCI_CONFIG2_READ_WRITE_OPCODE = 0x0C
-    EFI_BOOT_SCRIPT_TABLE_OPCODE = 0xAA
-
-
-#
-# Names of S3 Boot Script Opcodes
-#
-script_opcodes: Dict[int, str] = {
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_IO_WRITE_OPCODE: "S3_BOOTSCRIPT_IO_WRITE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_IO_READ_WRITE_OPCODE: "S3_BOOTSCRIPT_IO_READ_WRITE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_MEM_WRITE_OPCODE: "S3_BOOTSCRIPT_MEM_WRITE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_MEM_READ_WRITE_OPCODE: "S3_BOOTSCRIPT_MEM_READ_WRITE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_PCI_CONFIG_WRITE_OPCODE: "S3_BOOTSCRIPT_PCI_CONFIG_WRITE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_PCI_CONFIG_READ_WRITE_OPCODE: "S3_BOOTSCRIPT_PCI_CONFIG_READ_WRITE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_SMBUS_EXECUTE_OPCODE: "S3_BOOTSCRIPT_SMBUS_EXECUTE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_STALL_OPCODE: "S3_BOOTSCRIPT_STALL",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_DISPATCH_OPCODE: "S3_BOOTSCRIPT_DISPATCH",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_DISPATCH_2_OPCODE:             "S3_BOOTSCRIPT_DISPATCH_2",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_INFORMATION_OPCODE:            "S3_BOOTSCRIPT_INFORMATION",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_PCI_CONFIG2_WRITE_OPCODE:      "S3_BOOTSCRIPT_PCI_CONFIG2_WRITE",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_PCI_CONFIG2_READ_WRITE_OPCODE: "S3_BOOTSCRIPT_PCI_CONFIG2_READ_WRITE",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_IO_POLL_OPCODE:                "S3_BOOTSCRIPT_IO_POLL",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_MEM_POLL_OPCODE:               "S3_BOOTSCRIPT_MEM_POLL",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_PCI_CONFIG_POLL_OPCODE:        "S3_BOOTSCRIPT_PCI_CONFIG_POLL",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_PCI_CONFIG2_POLL_OPCODE:       "S3_BOOTSCRIPT_PCI_CONFIG2_POLL",
-    # S3BootScriptOpcode.EFI_BOOT_SCRIPT_TABLE_OPCODE:                  "S3_BOOTSCRIPT_TABLE",
-    S3BootScriptOpcode.EFI_BOOT_SCRIPT_TERMINATE_OPCODE: "S3_BOOTSCRIPT_TERMINATE"
-}
-
-
-# //*******************************************
-# // EFI_BOOT_SCRIPT_WIDTH
-# //*******************************************
-# typedef enum {
-# EfiBootScriptWidthUint8,
-# EfiBootScriptWidthUint16,
-# EfiBootScriptWidthUint32,
-# EfiBootScriptWidthUint64,
-# EfiBootScriptWidthFifoUint8,
-# EfiBootScriptWidthFifoUint16,
-# EfiBootScriptWidthFifoUint32,
-# EfiBootScriptWidthFifoUint64,
-# EfiBootScriptWidthFillUint8,
-# EfiBootScriptWidthFillUint16,
-# EfiBootScriptWidthFillUint32,
-# EfiBootScriptWidthFillUint64,
-# EfiBootScriptWidthMaximum
-# } EFI_BOOT_SCRIPT_WIDTH;
-
-class S3BootScriptWidth:
-    EFI_BOOT_SCRIPT_WIDTH_UINT8 = 0x00
-    EFI_BOOT_SCRIPT_WIDTH_UINT16 = 0x01
-    EFI_BOOT_SCRIPT_WIDTH_UINT32 = 0x02
-    EFI_BOOT_SCRIPT_WIDTH_UINT64 = 0x03
-
-
-script_width_sizes: Dict[int, int] = {
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT8: 1,
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT16: 2,
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT32: 4,
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT64: 8
-}
-
-script_width_values: Dict[int, int] = {
-    1: S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT8,
-    2: S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT16,
-    4: S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT32,
-    8: S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT64
-}
-
-script_width_formats: Dict[int, str] = {
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT8: 'B',
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT16: 'H',
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT32: 'I',
-    S3BootScriptWidth.EFI_BOOT_SCRIPT_WIDTH_UINT64: 'Q'
-}
-
-# //************************************************
-# // EFI_SMBUS_DEVICE_ADDRESS
-# //************************************************
-# typedef struct _EFI_SMBUS_DEVICE_ADDRESS {
-# UINTN SmbusDeviceAddress:7;
-# } EFI_SMBUS_DEVICE_ADDRESS;
-# //************************************************
-# // EFI_SMBUS_DEVICE_COMMAND
-# //************************************************
-# typedef UINTN EFI_SMBUS_DEVICE_COMMAND;
-#
-# //************************************************
-# // EFI_SMBUS_OPERATION
-# //************************************************
-# typedef enum _EFI_SMBUS_OPERATION {
-# EfiSmbusQuickRead,
-# EfiSmbusQuickWrite,
-# EfiSmbusReceiveByte,
-# EfiSmbusSendByte,
-# EfiSmbusReadByte,
-# EfiSmbusWriteByte,
-# EfiSmbusReadWord,
-# EfiSmbusWriteWord,
-# EfiSmbusReadBlock,
-# EfiSmbusWriteBlock,
-# EfiSmbusProcessCall,
-# EfiSmbusBWBRProcessCall
-# } EFI_SMBUS_OPERATION;
-
-
-class S3BootScriptSmbusOperation:
-    QUICK_READ = 0x00
-    QUICK_WRITE = 0x01
-    RECEIVE_BYTE = 0x02
-    SEND_BYTE = 0x03
-    READ_BYTE = 0x04
-    WRITE_BYTE = 0x05
-    READ_WORD = 0x06
-    WRITE_WORD = 0x07
-    READ_BLOCK = 0x08
-    WRITE_BLOCK = 0x09
-    PROCESS_CALL = 0x0A
-    BWBR_PROCESS_CALL = 0x0B
-
-
-class op_io_pci_mem:
-    def __init__(self, opcode: int, size: int, width: int, address: int, unknown: Optional[int], count: Optional[int], 
-                 buffer: Optional[bytes], value: Optional[int] = None, mask: Optional[int] = None):
-        self.opcode = opcode
-        self.size = size
-        self.width = width
-        self.address = address
-        self.unknown = unknown
-        self.count = count
-        self.value = value
-        self.mask = mask
-        self.name = script_opcodes[opcode]
-        self.buffer = buffer  # data[ self.size : ]
-        self.values = None
-        if self.count is not None and self.count > 0 and self.buffer is not None:
-            sz = self.count * script_width_sizes[self.width]
-            if len(self.buffer) != sz:
-                logger().log(f'[?] buffer size (0x{len(self.buffer):X}) != Width x Count (0x{sz:X})')
-            else:
-                self.values = list(struct.unpack((f'<{self.count:d}{script_width_formats[self.width]:1}'), self.buffer))
-
-    def __str__(self) -> str:
-        str_r = f'  Opcode : {self.name} (0x{self.opcode:04X})\n'
-        str_r += f'  Width  : 0x{self.width:02X} ({script_width_sizes[self.width]:X} bytes)\n'
-        str_r += f'  Address: 0x{self.address:08X}\n'
-        if self.value is not None:
-            str_r += f'  Value  : 0x{self.value:08X}\n'
-        if self.mask is not None:
-            str_r += f'  Mask   : 0x{self.mask:08X}\n'
-        if self.unknown is not None:
-            str_r += f'  Unknown: 0x{self.unknown:04X}\n'
-        if self.count is not None:
-            str_r += f'  Count  : 0x{self.count:X}\n'
-        if self.values is not None:
-            fmt = f'0x{{:0{script_width_sizes[self.width] * 2:d}X}}'
-            values_str = '  '.join([fmt.format(v) for v in self.values])
-            str_r += f'  Values : {values_str}\n'
-        elif self.buffer is not None:
-            str_r += f'  Buffer (size = 0x{len(self.buffer):X}):\n{dump_buffer(self.buffer, 16)}'
-        return str_r
-
-
-class op_smbus_execute:
-    def __init__(self, opcode: int, size: int, address: int, command: int, operation: int, peccheck: int):
-        self.opcode = opcode
-        self.size = size
-        self.address = address
-        self.command = command
-        self.operation = operation
-        self.peccheck = peccheck
-        self.name = script_opcodes[opcode]
-
-    def __str__(self) -> str:
-        str_r = f'  Opcode           : {self.name} (0x{self.opcode:04X})\n'
-        str_r += f'  Secondary Address: 0x{self.address:02X}\n'
-        str_r += f'  Command          : 0x{self.command:08X}\n'
-        str_r += f'  Operation        : 0x{self.operation:02X}\n'
-        str_r += f'  PEC Check        : {self.peccheck:d}\n'
-        return str_r
-
-# typedef struct {
-#  UINT16  OpCode;
-#  UINT8   Length;
-#  UINT64  Duration;
-# } EFI_BOOT_SCRIPT_STALL;
-
-
-class op_stall:
-    def __init__(self, opcode: int, size: int, duration: int):
-        self.opcode = opcode
-        self.size = size
-        self.duration = duration
-        self.name = script_opcodes[self.opcode]
-
-    def __str__(self) -> str:
-        str_r = f'  Opcode  : {self.name} (0x{self.opcode:04X})\n'
-        str_r += f'  Duration: 0x{self.duration:08X} (us)\n'
-        return str_r
-
-# typedef struct {
-#  UINT16                OpCode;
-#  UINT8                 Length;
-#  EFI_PHYSICAL_ADDRESS  EntryPoint;
-# } EFI_BOOT_SCRIPT_DISPATCH;
-
-
-class op_dispatch:
-    def __init__(self, opcode: int, size: int, entrypoint: int, context: Optional[int] = None):
-        self.opcode = opcode
-        self.size = size
-        self.entrypoint = entrypoint
-        self.context = context
-        self.name = script_opcodes[self.opcode]
-
-    def __str__(self) -> str:
-        str_r = f'  Opcode     : {self.name} (0x{self.opcode:04X})\n'
-        str_r += f'  Entry Point: 0x{self.entrypoint:016X}\n'
-        if self.context is not None:
-            str_r += f'  Context    : 0x{self.context:016X}\n'
-        return str_r
-
-# typedef struct {
-#  UINT16  OpCode;
-#  UINT8   Length;
-#  UINT32  Width;
-#  UINT64  Address;
-#  UINT64  Duration;
-#  UINT64  LoopTimes;
-# } EFI_BOOT_SCRIPT_MEM_POLL;
-
-
-class op_mem_poll:
-    def __init__(self, opcode: int, size: int, width: int, address: int, duration: int, looptimes: int):
-        self.opcode = opcode
-        self.size = size
-        self.width = width
-        self.address = address
-        self.duration = duration
-        self.looptimes = looptimes
-        self.name = 'S3_BOOTSCRIPT_MEM_POLL'
-
-    def __str__(self) -> str:
-        str_r = f'  Opcode    : {self.name} (0x{self.opcode:04X})\n'
-        str_r += f'  Width     : 0x{self.width:02X} ({script_width_sizes[self.width]:X} bytes)\n'
-        str_r += f'  Address   : 0x{self.address:016X}\n'
-        str_r += f'  Duration? : 0x{self.duration:016X}\n'
-        str_r += f'  LoopTimes?: 0x{self.looptimes:016X}\n'
-        return str_r
-
-
-class op_terminate:
-    def __init__(self, opcode: int, size: int):
-        self.opcode = opcode
-        self.size = size
-        self.name = script_opcodes[self.opcode]
-
-    def __str__(self) -> str:
-        return f'  Opcode     : {self.name} (0x{self.opcode:02X})\n'
-
-
-class op_unknown:
-    def __init__(self, opcode: int, size: int):
-        self.opcode = opcode
-        self.size = size
-
-    def __str__(self) -> str:
-        return f'  Opcode     : unknown (0x{self.opcode:02X})\n'
-
-
-class S3BOOTSCRIPT_ENTRY:
-    def __init__(self, script_type: int, index: Optional[int], offset_in_script: int, length: int, data: Optional[bytes] = None):
-        self.script_type = script_type
-        self.index = index
-        self.offset_in_script = offset_in_script
-        self.length = length
-        self.data = data
-        self.decoded_opcode = None
-        self.header_length = 0
-
-    def __str__(self) -> str:
-        entry_str = '' if self.index is None else (f'[{self.index:03d}] ')
-        entry_str += f'Entry at offset 0x{self.offset_in_script:04X} (len = 0x{self.length:X}, header len = 0x{self.header_length:X}):'
-        if self.data:
-            entry_str = entry_str + f'\nData:\n{dump_buffer_bytes(self.data, 16)}'
-        if self.decoded_opcode:
-            entry_str = entry_str + f'Decoded:\n{str(self.decoded_opcode)}'
-        return entry_str
 
 
 # #################################################################################################
