+++ chipsec/library/uefi/uefi_platform.py
@@ -27,11 +27,10 @@ from collections import namedtuple
 from uuid import UUID
 from typing import Dict, List, Tuple, Optional, Union, Any
 from chipsec.library import defines
-from chipsec.library.logger import logger
-from chipsec.library.uefi.uefi_common import bit_set, VARIABLE_SIGNATURE_VSS, S3BootScriptOpcode_MDE, op_io_pci_mem, S3BootScriptOpcode_EdkCompat, EFI_GUID_STR, EFI_GUID_SIZE
-from chipsec.library.uefi.uefi_common import op_stall, op_dispatch, op_terminate, op_mem_poll, op_unknown, get_3b_size, get_nvar_name, op_smbus_execute, script_width_formats
-from chipsec.library.uefi.uefi_common import S3BOOTSCRIPT_ENTRY, MAX_S3_BOOTSCRIPT_ENTRY_LENGTH, VARIABLE_STORE_FV_GUID, IS_VARIABLE_ATTRIBUTE, VARIABLE_DATA
-from chipsec.library.uefi.uefi_common import EFI_VARIABLE_BOOTSERVICE_ACCESS, EFI_VARIABLE_NON_VOLATILE, EFI_VARIABLE_RUNTIME_ACCESS, script_opcodes
+from chipsec.library.uefi.uefi_common import bit_set, VARIABLE_SIGNATURE_VSS, EFI_GUID_STR, EFI_GUID_SIZE
+from chipsec.library.uefi.uefi_common import get_3b_size, get_nvar_name
+from chipsec.library.uefi.uefi_common import VARIABLE_STORE_FV_GUID, IS_VARIABLE_ATTRIBUTE, VARIABLE_DATA
+from chipsec.library.uefi.uefi_common import EFI_VARIABLE_BOOTSERVICE_ACCESS, EFI_VARIABLE_NON_VOLATILE, EFI_VARIABLE_RUNTIME_ACCESS
 from chipsec.library.uefi.uefi_common import EFI_VARIABLE_HARDWARE_ERROR_RECORD, EFI_VARIABLE_AUTHENTICATED_WRITE_ACCESS, EFI_VARIABLE_TIME_BASED_AUTHENTICATED_WRITE_ACCESS
 from chipsec.library.uefi.uefi_fv import NextFwVolume, NextFwFile, EFI_FVB2_ERASE_POLARITY, EFI_FV_FILETYPE_RAW
 
@@ -917,377 +916,6 @@ def getEFIvariables_NtEnumerateSystemEnvironmentValuesEx2( nvram_buf ):
         return variables
 #    return ( start, next_var_offset, efi_var_buf, efi_var_hdr, efi_var_name, efi_var_data, guid_str(efi_var_hdr.guid0, efi_var_hdr.guid1, efi_var_hdr.guid2, efi_var_hdr.guid3), efi_var_hdr.Attributes )
 """
-
-
-#
-# Decoding S3 Resume Boot Script
-#
-
-class S3BootScriptType:
-    EFI_BOOT_SCRIPT_TYPE_DEFAULT = 0x00
-    EFI_BOOT_SCRIPT_TYPE_EDKCOMPAT = 0xAA
-
-
-def decode_s3bs_opcode(s3bootscript_type, script_data):
-    if S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_EDKCOMPAT == s3bootscript_type:
-        return decode_s3bs_opcode_edkcompat(script_data)
-    else:
-        return decode_s3bs_opcode_def(script_data)
-
-
-def encode_s3bs_opcode(s3bootscript_type: int, op: S3BOOTSCRIPT_ENTRY) -> bytes:
-    if S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_EDKCOMPAT == s3bootscript_type:
-        return encode_s3bs_opcode_edkcompat(op)
-    else:
-        return encode_s3bs_opcode_def(op)
-
-
-def decode_s3bs_opcode_def(data):
-    opcode = None
-    size = None
-    width = None
-    unknown = None
-    count = None
-    value = None
-    mask = None
-
-    op = None
-    opcode, = struct.unpack('<B', data[: 1])
-    try:
-        logger().log_hal(script_opcodes[opcode])
-    except:
-        pass
-    if S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_IO_WRITE_OPCODE == opcode:
-        frmt = '<BBHIQ'
-        size = struct.calcsize(frmt)
-        opcode, width, address, alignment, count = struct.unpack(frmt, data[: size])
-        op = op_io_pci_mem(opcode, size, width, address, unknown, count, data[size:], value, mask)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_IO_READ_WRITE_OPCODE == opcode:
-        frmt = '<BBHIQQ'
-        size = struct.calcsize(frmt)
-        opcode, width, address, _, value, mask = struct.unpack(frmt, data[: size])
-        op = op_io_pci_mem(opcode, size, width, address, unknown, count, None, value, mask)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_PCI_CONFIG_WRITE_OPCODE == opcode:
-        frmt = '<BBHIQQ'
-        size = struct.calcsize(frmt)
-        opcode, width, unknown, _, address, count = struct.unpack(frmt, data[: size])
-        op = op_io_pci_mem(opcode, size, width, address, unknown, count, data[size:], value, mask)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_PCI_CONFIG_READ_WRITE_OPCODE == opcode:
-        frmt = '<BBHIQQQ'
-        size = struct.calcsize(frmt)
-        opcode, width, unknown, _, address, value, mask = struct.unpack(frmt, data[: size])
-        op = op_io_pci_mem(opcode, size, width, address, unknown, count, None, value, mask)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_MEM_WRITE_OPCODE == opcode:
-        frmt = '<BBHIQQ'
-        size = struct.calcsize(frmt)
-        opcode, width, unknown, _, address, count = struct.unpack(frmt, data[: size])
-        op = op_io_pci_mem(opcode, size, width, address, unknown, count, data[size:], value, mask)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_MEM_READ_WRITE_OPCODE == opcode:
-        frmt = '<BBHIQQQ'
-        size = struct.calcsize(frmt)
-        opcode, width, unknown, _, address, value, mask = struct.unpack(frmt, data[: size])
-        op = op_io_pci_mem(opcode, size, width, address, unknown, count, None, value, mask)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_SMBUS_EXECUTE_OPCODE == opcode:
-        frmt = '<BBQBB'
-        size = struct.calcsize(frmt)
-        opcode, address, command, operation, peccheck = struct.unpack(frmt, data[: size])
-        op = op_smbus_execute(opcode, size, address, command, operation, peccheck)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_STALL_OPCODE == opcode:
-        frmt = '<BBQ'
-        size = struct.calcsize(frmt)
-        opcode, _, duration = struct.unpack(frmt, data[: size])
-        op = op_stall(opcode, size, duration)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_DISPATCH_OPCODE == opcode:
-        frmt = '<BBHIQ'
-        size = struct.calcsize(frmt)
-        opcode, _, _, _, entrypoint = struct.unpack(frmt, data[: size])
-        op = op_dispatch(opcode, size, entrypoint)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_DISPATCH_2_OPCODE == opcode:
-        frmt = '<BBHIQQ'
-        size = struct.calcsize(frmt)
-        opcode, _, _, _, entrypoint, context = struct.unpack(frmt, data[: size])
-        op = op_dispatch(opcode, size, entrypoint, context)
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_TERMINATE_OPCODE == opcode:
-        frmt = '<B'
-        size = struct.calcsize(frmt)
-        opcode, = struct.unpack(frmt, data[: size])
-        op = op_terminate(opcode, size)
-    else:
-        op = op_unknown(opcode, 1)
-        if logger().HAL:
-            logger().log_warning(f'Unrecognized opcode {opcode:X}')
-
-    return op
-
-#
-# @TODO: encode functions are not fully implemented
-#
-
-
-def encode_s3bs_opcode_def(op) -> bytes:
-    encoded_opcode = b''
-
-    if S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_IO_WRITE_OPCODE == op.opcode:
-        encoded_hdr = struct.pack('<BBHIQ', op.opcode, op.width, op.address, 0x0, op.count)
-        if op.values is None:
-            encoded_opcode = encoded_hdr + op.buffer
-        else:
-            encoded_opcode = encoded_hdr + struct.pack(script_width_formats[op.width] * op.count, *op.values)
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_IO_READ_WRITE_OPCODE == op.opcode:
-        encoded_opcode = struct.pack('<BBHIQQ', op.opcode, op.width, op.address, 0x0, op.value, op.mask)
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_PCI_CONFIG_WRITE_OPCODE == op.opcode or \
-            S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_MEM_WRITE_OPCODE == op.opcode:
-        encoded_hdr = struct.pack('<BBHIQQ', op.opcode, op.width, op.unknown, 0x0, op.address, op.count)
-        if op.values is None:
-            encoded_opcode = encoded_hdr + op.buffer
-        else:
-            encoded_opcode = encoded_hdr + struct.pack(script_width_formats[op.width] * op.count, *op.values)
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_PCI_CONFIG_READ_WRITE_OPCODE == op.opcode:
-        frmt = '<BBHIQQQ'
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_MEM_READ_WRITE_OPCODE == op.opcode:
-        encoded_opcode = struct.pack('<BBHIQQQ', op.opcode, op.width, op.unknown, 0x0, op.address, op.value, op.mask)
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_SMBUS_EXECUTE_OPCODE == op.opcode:
-        frmt = '<BBQBB'
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_STALL_OPCODE == op.opcode:
-        frmt = '<BBQ'
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_DISPATCH_OPCODE == op.opcode:
-        encoded_opcode = struct.pack('<BBHIQ', op.opcode, 0x0, 0x0, 0x0, op.entrypoint)
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_DISPATCH_2_OPCODE == op.opcode:
-        encoded_opcode = struct.pack('<BBHIQQ', op.opcode, 0x0, 0x0, 0x0, op.entrypoint, op.context)
-
-    elif S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_TERMINATE_OPCODE == op.opcode:
-        frmt = '<B'
-
-    else:
-        if logger().HAL:
-            logger().log_warning(f'Unrecognized opcode {op.opcode:X}')
-
-    return encoded_opcode
-
-
-def decode_s3bs_opcode_edkcompat(data: bytes):
-    opcode = None
-    width = None
-    count = None
-    value = None
-    mask = None
-
-    op = None
-
-    hdr_frmt = '<HB'
-    header_size = struct.calcsize(hdr_frmt)
-    opcode, size = struct.unpack(hdr_frmt, data[: header_size])
-    opcode_data = data[header_size:]
-    try:
-        logger().log_hal(script_opcodes[opcode])
-    except:
-        pass
-
-    if S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_IO_WRITE_OPCODE == opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_PCI_CONFIG_WRITE_OPCODE == opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_MEM_WRITE_OPCODE == opcode:
-
-        frmt = '<IIQ'
-        op_size = struct.calcsize(frmt)
-        width, count, address = struct.unpack(frmt, opcode_data[: op_size])
-        op = op_io_pci_mem(opcode, size, width, address, None, count, opcode_data[op_size:], value, mask)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_IO_READ_WRITE_OPCODE == opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_PCI_CONFIG_READ_WRITE_OPCODE == opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_MEM_READ_WRITE_OPCODE == opcode:
-        frmt = '<IQ'
-        sz = struct.calcsize(frmt)
-        width, address = struct.unpack(frmt, opcode_data[: sz])
-        frmt = 2 * script_width_formats[width]
-        op_size = sz + struct.calcsize(frmt)
-        value, mask = struct.unpack(frmt, opcode_data[sz: op_size])
-        op = op_io_pci_mem(opcode, size, width, address, None, count, None, value, mask)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_SMBUS_EXECUTE_OPCODE == opcode:
-        if logger().UTIL_TRACE or logger().HAL:
-            logger().log_warning(f'Cannot parse opcode {opcode:X} yet')
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_STALL_OPCODE == opcode:
-        frmt = '<Q'
-        op_size = struct.calcsize(frmt)
-        duration, = struct.unpack(frmt, opcode_data[: op_size])
-        op = op_stall(opcode, size, duration)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_DISPATCH_OPCODE == opcode:
-        frmt = '<Q'
-        op_size = struct.calcsize(frmt)
-        entrypoint, = struct.unpack(frmt, opcode_data[: op_size])
-        op = op_dispatch(opcode, size, entrypoint)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_MEM_POLL_OPCODE == opcode:
-        frmt = '<IQQQ'
-        op_size = struct.calcsize(frmt)
-        width, address, duration, looptimes = struct.unpack(frmt, opcode_data[: op_size])
-        op = op_mem_poll(opcode, size, width, address, duration, looptimes)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_TERMINATE_OPCODE == opcode:
-        op = op_terminate(opcode, size)
-
-    else:
-        op = op_unknown(opcode, size)
-        if logger().HAL:
-            logger().log_warning(f'Unrecognized opcode {opcode:X}')
-
-    return op
-
-#
-# @TODO: encode functions are not fully implemented
-#
-
-
-def encode_s3bs_opcode_edkcompat(op: S3BOOTSCRIPT_ENTRY) -> bytes:
-    encoded_opcode = b''
-
-    if S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_IO_WRITE_OPCODE == op.opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_PCI_CONFIG_WRITE_OPCODE == op.opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_MEM_WRITE_OPCODE == op.opcode:
-
-        encoded_hdr = struct.pack('<IIQ', op.width, op.count, op.address)
-        if op.values is None:
-            encoded_opcode = encoded_hdr + op.buffer
-        else:
-            encoded_opcode = encoded_hdr + struct.pack(script_width_formats[op.width] * op.count, *op.values)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_IO_READ_WRITE_OPCODE == op.opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_PCI_CONFIG_READ_WRITE_OPCODE == op.opcode or \
-            S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_MEM_READ_WRITE_OPCODE == op.opcode:
-
-        frmt = f'<IQ2{script_width_formats[op.width]}'
-        encoded_opcode = struct.pack(frmt, op.width, op.address, op.value, op.mask)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_SMBUS_EXECUTE_OPCODE == op.opcode:
-        pass
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_STALL_OPCODE == op.opcode:
-        frmt = '<Q'
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_DISPATCH_OPCODE == op.opcode:
-        encoded_opcode = struct.pack('<Q', op.entrypoint)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_MEM_POLL_OPCODE == op.opcode:
-        encoded_opcode = struct.pack('<IQQQ', op.width, op.address, op.duration, op.looptimes)
-
-    elif S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_TERMINATE_OPCODE == op.opcode:
-        pass
-
-    return encoded_opcode
-
-
-def parse_s3bootscript_entry(s3bootscript_type: int, script: bytes, off: int, log_script: bool = False):
-    entry_index = None
-    entry_length = 0
-    opcode = None
-    entry_data = None
-
-    remaining_len = len(script[off:])
-
-    if S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_EDKCOMPAT == s3bootscript_type:
-        fhdr = '<HB'
-        hdr_length = struct.calcsize(fhdr)
-        if remaining_len < hdr_length:
-            if logger().HAL:
-                logger().log_warning(f'The script should have at least 0x{hdr_length:X} bytes to parse next entry')
-            return (0, None)
-
-        opcode, entry_length = struct.unpack(fhdr, script[off: off + hdr_length])
-        if S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_TERMINATE_OPCODE == opcode:
-            entry_length = hdr_length
-        entry_data = script[off: off + entry_length]
-
-        if entry_length > MAX_S3_BOOTSCRIPT_ENTRY_LENGTH:
-            logger().log_error(f'[uefi] Unrecognized S3 boot script format (entry length = 0x{entry_length:X})')
-            return (0, None)
-
-        s3script_entry = S3BOOTSCRIPT_ENTRY(s3bootscript_type, entry_index, off, entry_length, entry_data)
-
-    else:  # S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_DEFAULT
-
-        fhdr = '<II'
-        hdr_length = struct.calcsize(fhdr)
-        f = fhdr + 'B'
-        if remaining_len < (hdr_length + 1):
-            if logger().HAL:
-                logger().log_warning(f'The script should have at least 0x{hdr_length + 1:X} bytes to parse next entry')
-            return (0, None)
-
-        entry_index, entry_length, opcode = struct.unpack(f, script[off: off + hdr_length + 1])
-        if S3BootScriptOpcode_MDE.EFI_BOOT_SCRIPT_TERMINATE_OPCODE == opcode:
-            entry_length = hdr_length + 1
-            entry_index = -1
-        entry_data = script[off + hdr_length: off + entry_length]
-
-        if entry_length > MAX_S3_BOOTSCRIPT_ENTRY_LENGTH:
-            logger().log_error(f'[uefi] Unrecognized S3 boot script format (entry length = 0x{entry_length:X})')
-            return (0, None)
-
-        s3script_entry = S3BOOTSCRIPT_ENTRY(s3bootscript_type, entry_index, off, entry_length, entry_data)
-        s3script_entry.header_length = hdr_length
-
-    s3script_entry.decoded_opcode = decode_s3bs_opcode(s3bootscript_type, s3script_entry.data)
-
-    if log_script:
-        logger().log(str(s3script_entry))
-    return (opcode, s3script_entry)
-
-
-def encode_s3bootscript_entry(entry) -> Optional[bytes]:
-    if S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_EDKCOMPAT == entry.script_type:
-        entry_hdr_buf = struct.pack('<HB', entry.decoded_opcode.opcode, entry.length)
-    else:  # S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_DEFAULT
-        entry_hdr_buf = struct.pack('<II', entry.index, entry.length)
-
-    entry_val_buf = encode_s3bs_opcode(entry.script_type, entry.decoded_opcode)
-    entry_buf = None
-    if entry_val_buf is not None:
-        entry_buf = entry_hdr_buf + entry_val_buf
-    else:
-        logger().log_warning(f'Could not encode opcode of boot script entry (type 0x{entry.script_type:X})')
-
-    return entry_buf
-
-
-def create_s3bootscript_entry_buffer(script_type: int, op, index=None) -> bytes:
-    entry_val_buf = encode_s3bs_opcode(script_type, op)
-    length = len(entry_val_buf)
-    if S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_EDKCOMPAT == script_type:
-        length += struct.calcsize('<HB')
-        entry_hdr_buf = struct.pack('<HB', op.opcode, length)
-    else:  # S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_DEFAULT
-        length += struct.calcsize('<II')
-        entry_hdr_buf = struct.pack('<II', index, length)
-
-    return (entry_hdr_buf + entry_val_buf)
-
-
-def id_s3bootscript_type(script: bytes, log_script: bool = False) -> Tuple[int, int]:
-    script_header_length = 0
-
-    start_op, = struct.unpack('<B', script[: 1])
-    if S3BootScriptOpcode_EdkCompat.EFI_BOOT_SCRIPT_TABLE_OPCODE == start_op:
-        logger().log_hal('S3 Boot Script AA Parser')
-        script_type = S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_EDKCOMPAT
-        if log_script:
-            logger().log(f'[uefi] Start opcode 0x{start_op:X}')
-        # MdeModulePkg\Library\PiDxeS3BootScriptLib\BootScriptInternalFormat.h
-        script_header_length = struct.calcsize("<HBHLHH")
-    else:
-        logger().log_hal('S3 Boot Script DEFAULT Parser')
-        script_type = S3BootScriptType.EFI_BOOT_SCRIPT_TYPE_DEFAULT
-
-    return (script_type, script_header_length)
 
 
 #
