+++ chipsec/modules/common/uefi/s3bootscript.py
@@ -53,8 +53,8 @@ Examples:
 from chipsec.module_common import BaseModule, MTAG_BIOS, MTAG_SMM, MTAG_SECUREBOOT
 from chipsec.library.returncode import ModuleResult
 from chipsec.library.defines import BOUNDARY_1MB, BOUNDARY_4GB
-from chipsec.hal.common.uefi import UEFI, parse_script
-from chipsec.library.uefi.uefi_common import S3BootScriptOpcode, S3BOOTSCRIPT_ENTRY
+from chipsec.hal.common.uefi import UEFI
+from chipsec.library.uefi.sleep_states import S3BootScriptOpcode, S3BOOTSCRIPT_ENTRY, parse_script
 from typing import List
 
 TAGS = [MTAG_BIOS, MTAG_SMM, MTAG_SECUREBOOT]
